safari-600.7-branch/LayoutTests/fast/canvas/canvas-with-illegal-args-expected.txt

Series of tests to ensure correct behaviour of calling canvas methods with illegal arguments (Infintiy and NaN)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test scale.
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
Test translate.
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
Test rotate.
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
